+++ packages/vscode-extension/src/commands/build.ts
@@ -10,7 +10,7 @@ import {
 import { cli } from "../cli";
 import { ensureProjectList, projects, projectTargetList } from "../project";
 
-export function cmdBuildFile(runInTerminal: boolean = false) {
+export async function cmdBuildFile(runInTerminal: boolean = false) {
   // get the active text editor's file name
   let activeEditor = window.activeTextEditor;
   if (!activeEditor) {
@@ -27,17 +27,47 @@ export function cmdBuildFile(runInTerminal: boolean = false) {
   }
   let fileName = activeEditor.document.fileName;
   console.log(`Filename:`, fileName);
-  if (runInTerminal) {
-    let terminal = utils.setupTerminal(pretextTerminal);
-    terminal.sendText("pretext build " + fileName);
-  } else {
-    runPretext(cli.cmd(), "build", fileName);
-  }
+  await ensureProjectList();
+  let targetSelection = projectTargetList({ standalones: true });
+  window.showQuickPick(targetSelection).then((qpSelection) => {
+    if (!qpSelection) {
+      return;
+    }
+    if (runInTerminal) {
+      let terminal = utils.setupTerminal(
+        pretextTerminal,
+        qpSelection.description
+      );
+      terminal.sendText(
+        "pretext build " + qpSelection.label + " -i " + fileName
+      );
+    } else {
+      runPretext(
+        cli.cmd(),
+        "build",
+        qpSelection.label + " -i " + fileName,
+        qpSelection.description
+      );
+    }
+    updateLastTarget({
+      name: qpSelection.label,
+      path: qpSelection.description,
+      standalone: true,
+      filename: fileName,
+    });
+    setTopCommand("Build current file as " + lastTarget.name);
+  });
+  //if (runInTerminal) {
+  //  let terminal = utils.setupTerminal(pretextTerminal);
+  //  terminal.sendText("pretext build " + fileName);
+  //} else {
+  //  runPretext(cli.cmd(), "build", fileName);
+  //}
 }
 
 export async function cmdBuildAny(runInTerminal: boolean = false) {
   await ensureProjectList();
-  let targetSelection = projectTargetList();
+  let targetSelection = projectTargetList({ standalones: false });
   // Show choice dialog and pass correct command to runPretext based on selection.
   window.showQuickPick(targetSelection).then((qpSelection) => {
     if (!qpSelection) {
@@ -60,8 +90,10 @@ export async function cmdBuildAny(runInTerminal: boolean = false) {
     updateLastTarget({
       name: qpSelection.label,
       path: qpSelection.description,
+      standalone: false,
+      filename: "",
     });
-    setTopCommand("Build " + lastTarget.name);
+    setTopCommand("Build target: " + lastTarget.name);
   });
 }
 
@@ -71,19 +103,28 @@ export async function cmdBuildLast(runInTerminal: boolean = false) {
     let terminal = utils.setupTerminal(pretextTerminal, projects[0].root);
     terminal.sendText("pretext build");
   } else {
-    let projectPath = lastTarget.path;
-    if (projectPath === "") {
-      //Fix project path if this is the first time default target is called.
-      projectPath = projects[0].root;
-      console.log("projectPath: ", projectPath);
+    if (lastTarget.standalone) {
+      runPretext(
+        cli.cmd(),
+        "build",
+        lastTarget.name + " -i " + lastTarget.filename,
+        lastTarget.path
+      );
+    } else {
+      let projectPath = lastTarget.path;
+      if (projectPath === "") {
+        //Fix project path if this is the first time default target is called.
+        projectPath = projects[0].root;
+        console.log("projectPath: ", projectPath);
+      }
+      runPretext(cli.cmd(), "build", lastTarget.name, projectPath);
     }
-    runPretext(cli.cmd(), "build", lastTarget.name, projectPath);
   }
 }
 
 export async function cmdGenerate(runInTerminal: boolean = false) {
   await ensureProjectList();
-  let targetSelection = projectTargetList();
+  let targetSelection = projectTargetList({ standalones: false });
   // Show choice dialog and pass correct command to runPretext based on selection.
   window.showQuickPick(targetSelection).then((qpSelection) => {
     if (!qpSelection) {